+++ packages/dialtone-vue2/components/button/button.vue
@@ -192,7 +192,7 @@ export default {
 
     /**
      * The color of the button.
-     * @values default, muted, danger, inverted
+     * @values default, muted, danger, positive, inverted
      */
     kind: {
       type: String,