+++ packages/dialtone-vue3/components/checkbox/checkbox.vue
@@ -27,12 +27,12 @@
       </div>
     </label>
     <div
-      v-if="hasDescriptionOrMessages"
+      v-if="$slots.description || description || hasMessages"
       class="d-checkbox__messages"
       data-qa="checkbox-description-messages"
     >
       <div
-        v-if="hasDescription"
+        v-if="$slots.description || description"
         :class="['d-description', descriptionClass]"
         v-bind="descriptionChildProps"
         data-qa="checkbox-description"
@@ -88,6 +88,13 @@ export default {
      * @type {Boolean}
      */
     'input',
+    /**
+     * Event fired to sync the modelValue prop with the parent component
+     *
+     * @event update:modelValue
+     * @type {Boolean}
+     */
+    'update:modelValue',
 
     /**
      * Native input focusin event
@@ -117,14 +124,6 @@ export default {
 
     hasLabel () {
       return !!(this.$slots.default || this.label);
-    },
-
-    hasDescription () {
-      return !!(this.$slots.description || this.description);
-    },
-
-    hasDescriptionOrMessages () {
-      return this.hasDescription || this.hasMessages;
     },
 
     hasMessages () {
@@ -180,6 +179,7 @@ export default {
 
       // emit the state of the checkbox
       this.$emit('input', checked);
+      this.$emit('update:modelValue', checked);
     },
 
     runValidations () {