+++ packages/dialtone-vue2/CHANGELOG.md
@@ -1,3 +1,42 @@
+# [2.186.0](https://github.com/dialpad/dialtone/compare/dialtone-vue2/v2.185.0...dialtone-vue2/v2.186.0) (2025-04-25)
+
+
+### Features
+
+* DLT-2307 dialpad rebrand 2025 ([#733](https://github.com/dialpad/dialtone/issues/733)) ([1be97a2](https://github.com/dialpad/dialtone/commit/1be97a2a6c7ac81b85e7efd3b0bef57fcf54f329))
+
+# [2.185.0](https://github.com/dialpad/dialtone/compare/dialtone-vue2/v2.184.1...dialtone-vue2/v2.185.0) (2025-04-24)
+
+
+### Bug Fixes
+
+* **Callbar Button With Dropdown:** DLT-2441 add modal false ([#723](https://github.com/dialpad/dialtone/issues/723)) ([b031e78](https://github.com/dialpad/dialtone/commit/b031e785b698b45b6cc85ab4a52f1b029ed1854f))
+
+
+### Features
+
+* **Breakpoints, Stack, Layout:** DLT-2462 reverse default breakpoints ([#722](https://github.com/dialpad/dialtone/issues/722)) ([b4c01e1](https://github.com/dialpad/dialtone/commit/b4c01e1929c1c895796fc3859295e9f6674db1a0))
+
+## [2.184.1](https://github.com/dialpad/dialtone/compare/dialtone-vue2/v2.184.0...dialtone-vue2/v2.184.1) (2025-04-21)
+
+
+### Bug Fixes
+
+* **Checkbox, Radio:** NO-JIRA description slot not being reactive ([#725](https://github.com/dialpad/dialtone/issues/725)) ([7a29e67](https://github.com/dialpad/dialtone/commit/7a29e67fd3f8efb9b4036dd98a716903aa7ad3d1))
+
+# [2.184.0](https://github.com/dialpad/dialtone/compare/dialtone-vue2/v2.183.2...dialtone-vue2/v2.184.0) (2025-04-17)
+
+
+### Bug Fixes
+
+* **Localization:** NO-JIRA remove localization from popover ([#719](https://github.com/dialpad/dialtone/issues/719)) ([ed5da5c](https://github.com/dialpad/dialtone/commit/ed5da5c67946354685d95db5f4ac9ac470605f88))
+* NO-JIRA bundle issues ([#715](https://github.com/dialpad/dialtone/issues/715)) ([ae16776](https://github.com/dialpad/dialtone/commit/ae16776ee2d8128dae891418a41a84075e7fa0c0))
+
+
+### Features
+
+* **Toast:** DLT-2418 support multiple layouts within toast ([#678](https://github.com/dialpad/dialtone/issues/678)) ([c593554](https://github.com/dialpad/dialtone/commit/c593554cc59fc171a4ce42e670c539a83d64a06c))
+
 ## [2.183.2](https://github.com/dialpad/dialtone/compare/dialtone-vue2/v2.183.1...dialtone-vue2/v2.183.2) (2025-04-16)